# Optimize backward of log2 and log10