Slightly optimized backward computation of `F.log2` and `F.log10`.

Time per single `F.log2.backward` call:

CPU:
```
BEFORE: 78.110 us   +/- 3.850 us
AFTER:  70.842 us   +/- 2.635 us
```

GPU:
```
BEFORE: 192.078 us   +/- 6.653 us
AFTER:  166.129 us   +/- 7.813 us
```